implement printfromdataurl util
